Update removeAt function
Previous removeAt function would throw a type error if the index is equals to the length of the list. This is because the check at line 184 checked if the index is greater or equals to the length of the list whereas it should only check if the index is greater than the length and if true, it should throw a type error.