solves issue #1280
traverseDFS() is not working properly ,

while logging it says  [ undefined ].

and there is also an issue with searchDFS() function,

the following issue i have faced

The stack array is being used as a queue instead of a stack. A queue follows the first-in, first-out (FIFO) principle, while a stack follows the last-in, first-out (LIFO) principle. For a DFS algorithm, we need to use a stack data structure.

The code is pushing elements to the stack using the tree[node.right] and tree[node.left] expressions, but it should be using node.right and node.left directly. The tree array should not be used here.

The code is using tree[0] as the starting node for the search, but it should be using the entire tree object as the starting node.